requirements.txt
requests
selenium
beautifulsoup4
python-dotenv
Pillow
XlsxWriter